# Graphemes: add fast-path for printable-ASCII