@rivo 2022-07-22
I'm wondering why you would use this package if you only deal with printable ASCII characters.