@dchapes 2022-07-23
@charlievieth again, as with my previous comment, if this package uses an internal `string` representation it should just ditch the `[]rune` conversion entirely.