@rivo 2023-07-22
I prefer not to add these kind of exceptions.